Refactor README for clarity and structure; add key directories and files section. Update detect_disease.py to streamline prediction result logging. Modify train_diseases.py to enhance dataset loading and label encoding process.

+++ train_diseases.py
@@ -11,15 +11,14 @@ from sklearn.preprocessing import LabelEncoder
 import os
 import json
 
-# === Load dataset ===
-df = pd.read_csv("Symptom2Disease.csv")  # Replace with your CSV file
-df["index"] = df["index"].astype(int)  # Ensure index is int
+df = pd.read_csv("Symptom2Disease.csv")
+df["index"] = df["index"].astype(int)
 
-# === Encode labels ===
+
 label_encoder = LabelEncoder()
 df["label_id"] = label_encoder.fit_transform(df["label"])
 
-# Save label map
+
 label2id = {label: int(idx) for idx, label in enumerate(label_encoder.classes_)}
 id2label = {v: k for k, v in label2id.items()}
 
@@ -27,12 +26,10 @@ os.makedirs("disease_classifier", exist_ok=True)
 with open("disease_classifier/label_map.json", "w") as f:
     json.dump(label2id, f)
 
-# === Split based on index % 7 ===
 train_df = df[df["index"] % 7 != 0].reset_index(drop=True)
 
 print(f"Training samples: {len(train_df)}")
 
-# === Load tokenizer and model ===
 model_name = "emilyalsentzer/Bio_ClinicalBERT"
 tokenizer = AutoTokenizer.from_pretrained(model_name)
 model = AutoModelForSequenceClassification.from_pretrained(
@@ -42,7 +39,6 @@ model = AutoModelForSequenceClassification.from_pretrained(
     id2label=id2label
 )
 
-# === Dataset class ===
 class DiseaseDataset(Dataset):
     def __init__(self, texts, labels):
         self.encodings = tokenizer(texts, truncation=True, padding=True, max_length=512)
@@ -56,9 +52,9 @@ class DiseaseDataset(Dataset):
     def __len__(self):
         return len(self.labels)
 
-# === Create datasets ===
+
 train_dataset = DiseaseDataset(train_df["text"].tolist(), train_df["label_id"].tolist())
-# === Training arguments ===
+
 training_args = TrainingArguments(
     output_dir="./results",
     num_train_epochs=3,
@@ -71,16 +67,13 @@ training_args = TrainingArguments(
 )
 
 
-# === Trainer setup ===
 trainer = Trainer(
     model=model,
     args=training_args,
     train_dataset=train_dataset
 )
 
-# === Train! ===
 trainer.train()
 
-# === Save model and tokenizer ===
 model.save_pretrained("disease_classifier")
 tokenizer.save_pretrained("disease_classifier")
